fix: Prevent Duplicate onClick and href triggers
for components wrapping uncloned children, only trigger the parent onClick function when hotkeys are activated or when the onClick trigger is native to the parent. When hotkeys are activated, children click events are disabled by setting 'pointerEvents' to 'none' on an inner div also wrapping children. When hotkeys are not activated, this styling is disabled to allow for event propogation from children. FIX: initialize 'btnFound' var to 'false' to fix bug in which the components wrapped in a button on first render would be identified  as buttons on the 2nd render, resulting in btnFound being set to true.